requirements.txt
Flask
Flask-SocketIO
langchain-groq
python-dotenv
gunicorn
eventlet
httpx==0.27.0